+++ src/api/models/view_models/habitatcomplexity.py
@@ -49,9 +49,7 @@ FROM
     sample_unit_id = models.UUIDField()
     sample_time = models.TimeField()
     transect_number = models.PositiveSmallIntegerField()
-    transect_len_surveyed = models.PositiveSmallIntegerField(
-        verbose_name=_("transect length surveyed (m)")
-    )
+    transect_len_surveyed = models.DecimalField(max_digits=4, decimal_places=1, verbose_name=_("transect length surveyed (m)"))
     reef_slope = models.CharField(max_length=50)
     interval_size = models.DecimalField(
         max_digits=4, decimal_places=2, default=0.5, verbose_name=_("interval size (m)")
@@ -130,9 +128,7 @@ ON (habcomp_su.pseudosu_id = habcomp_obs.pseudosu_id);
 
     sample_unit_ids = JSONField()
     transect_number = models.PositiveSmallIntegerField()
-    transect_len_surveyed = models.PositiveSmallIntegerField(
-        verbose_name=_("transect length surveyed (m)")
-    )
+    transect_len_surveyed = models.DecimalField(max_digits=4, decimal_places=1, verbose_name=_("transect length surveyed (m)"))
     reef_slope = models.CharField(max_length=50)
     score_avg = models.DecimalField(decimal_places=2, max_digits=3)
     data_policy_habitatcomplexity = models.CharField(max_length=50)
